secure/storage/src/config.rs -> config/src/config/secure_storage_config.rs
