+++ config/Cargo.toml
@@ -22,8 +22,10 @@ thiserror = "1.0.20"
 lcs = { path = "../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 libra-crypto-derive = { path = "../crypto/crypto-derive", version = "0.1.0" }
+libra-global-constants = { path = "../config/global-constants", version = "0.1.0"}
 libra-logger = { path = "../common/logger", version = "0.1.0" }
 libra-network-address = { path = "../network/network-address", version = "0.1.0" }
+libra-secure-storage = { path = "../secure/storage", version = "0.1.0" }
 libra-temppath = { path = "../common/temppath", version = "0.1.0" }
 libra-types = { path = "../types", version = "0.1.0" }
 libra-workspace-hack = { path = "../common/workspace-hack", version = "0.1.0" }
