+++ config/src/config/secure_storage_config.rs
@@ -1,9 +1,12 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::{CryptoStorage, KVStorage, Storage};
-use libra_config::config::{Identity, NetworkConfig, WaypointConfig};
+use crate::config::{Identity, NetworkConfig, SecureBackend, WaypointConfig};
 use libra_crypto::x25519;
+use libra_secure_storage::{
+    CryptoStorage, GitHubStorage, InMemoryStorage, KVStorage, NamespacedStorage, OnDiskStorage,
+    Storage, VaultStorage,
+};
 use libra_types::{waypoint::Waypoint, PeerId};
 use std::{convert::TryInto, str::FromStr};
 
@@ -58,4 +61,41 @@ pub fn waypoint(config: &WaypointConfig) -> Waypoint {
         WaypointConfig::None => None,
     };
     waypoint.expect("waypoint should be present")
+}
+
+impl From<&SecureBackend> for Storage {
+    fn from(backend: &SecureBackend) -> Self {
+        match backend {
+            SecureBackend::GitHub(config) => {
+                let storage = GitHubStorage::new(
+                    config.repository_owner.clone(),
+                    config.repository.clone(),
+                    config.token.read_token().expect("Unable to read token"),
+                );
+                if let Some(namespace) = &config.namespace {
+                    Storage::from(NamespacedStorage::new(Box::new(storage), namespace.clone()))
+                } else {
+                    Storage::from(storage)
+                }
+            }
+            SecureBackend::InMemoryStorage => Storage::from(InMemoryStorage::new()),
+            SecureBackend::OnDiskStorage(config) => {
+                let storage = OnDiskStorage::new(config.path());
+                if let Some(namespace) = &config.namespace {
+                    Storage::from(NamespacedStorage::new(Box::new(storage), namespace.clone()))
+                } else {
+                    Storage::from(storage)
+                }
+            }
+            SecureBackend::Vault(config) => Storage::from(VaultStorage::new(
+                config.server.clone(),
+                config.token.read_token().expect("Unable to read token"),
+                config.namespace.clone(),
+                config
+                    .ca_certificate
+                    .as_ref()
+                    .map(|_| config.ca_certificate().unwrap()),
+            )),
+        }
+    }
 }
