+++ secure/storage/Cargo.toml
@@ -19,7 +19,6 @@ serde_json = "1.0.55"
 thiserror = "1.0.20"
 
 lcs = { path = "../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
-libra-config = { path = "../../config", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0" }
 libra-global-constants = { path = "../../config/global-constants", version = "0.1.0"}
 libra-secure-time = { path = "../time", version = "0.1.0" }
