+++ secure/storage/src/lib.rs
@@ -3,10 +3,6 @@
 
 #![forbid(unsafe_code)]
 
-use libra_config::config::SecureBackend;
-use std::convert::From;
-
-pub mod config;
 mod crypto_kv_storage;
 mod crypto_storage;
 mod error;
@@ -34,43 +30,6 @@ pub use crate::{
     value::Value,
     vault::VaultStorage,
 };
-
-impl From<&SecureBackend> for Storage {
-    fn from(backend: &SecureBackend) -> Self {
-        match backend {
-            SecureBackend::GitHub(config) => {
-                let storage = GitHubStorage::new(
-                    config.repository_owner.clone(),
-                    config.repository.clone(),
-                    config.token.read_token().expect("Unable to read token"),
-                );
-                if let Some(namespace) = &config.namespace {
-                    Storage::from(NamespacedStorage::new(Box::new(storage), namespace.clone()))
-                } else {
-                    Storage::from(storage)
-                }
-            }
-            SecureBackend::InMemoryStorage => Storage::from(InMemoryStorage::new()),
-            SecureBackend::OnDiskStorage(config) => {
-                let storage = OnDiskStorage::new(config.path());
-                if let Some(namespace) = &config.namespace {
-                    Storage::from(NamespacedStorage::new(Box::new(storage), namespace.clone()))
-                } else {
-                    Storage::from(storage)
-                }
-            }
-            SecureBackend::Vault(config) => Storage::from(VaultStorage::new(
-                config.server.clone(),
-                config.token.read_token().expect("Unable to read token"),
-                config.namespace.clone(),
-                config
-                    .ca_certificate
-                    .as_ref()
-                    .map(|_| config.ca_certificate().unwrap()),
-            )),
-        }
-    }
-}
 
 #[cfg(test)]
 mod tests;
